Maybe

+++ 2016-20.go
@@ -74,7 +74,7 @@ func getIps(ctx context.Context, data string, m int64, rlog func(context.Context
 
 	sumv := int64(0)
 	for i := 0; i < len(ranges)-1; i++ {
-		sumv += ranges[i+1][0] - ranges[i][1] - 1
+		sumv += ranges[i+1][0] - ranges[i][1] - int64(1)
 	}
 
 	sumv += m - ranges[len(ranges)-1][1]
